move metric type setter to GoogleCredentials and metrics enum to get metric changes within oauth2_http package.
