--- credentials/java/com/google/auth/CredentialTypeForMetrics.java
@@ -1,64 +1,0 @@
-/*
- * Copyright 2024 Google LLC
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are
- * met:
- *
- *    * Redistributions of source code must retain the above copyright
- * notice, this list of conditions and the following disclaimer.
- *    * Redistributions in binary form must reproduce the above
- * copyright notice, this list of conditions and the following disclaimer
- * in the documentation and/or other materials provided with the
- * distribution.
- *
- *    * Neither the name of Google LLC nor the names of its
- * contributors may be used to endorse or promote products derived from
- * this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package com.google.auth;
-
-/**
- * Defines the different types of credentials that can be used for metrics.
- *
- * <p>Each credential type is associated with a label that is used for reporting purposes. Add new
- * enum constant only when corresponding configs established.
- *
- * <p>Credentials with type {@code CredentialTypeForMetrics.DO_NOT_SEND} is default value for
- * credential implementations that do not set type specifically. It is not expected to send metrics.
- *
- * <p>
- *
- * @see #getLabel()
- */
-public enum CredentialTypeForMetrics {
-  USER_CREDENTIALS("u"),
-  SERVICE_ACCOUNT_CREDENTIALS_AT("sa"),
-  SERVICE_ACCOUNT_CREDENTIALS_JWT("jwt"),
-  VM_CREDENTIALS("mds"),
-  IMPERSONATED_CREDENTIALS("imp"),
-  DO_NOT_SEND("dns");
-
-  private final String label;
-
-  private CredentialTypeForMetrics(String label) {
-    this.label = label;
-  }
-
-  public String getLabel() {
-    return label;
-  }
-}
