+++ oauth2_http/java/com/google/auth/oauth2/ImpersonatedCredentials.java
@@ -43,7 +43,7 @@ import com.google.api.client.http.HttpTransport;
 import com.google.api.client.http.json.JsonHttpContent;
 import com.google.api.client.json.JsonObjectParser;
 import com.google.api.client.util.GenericData;
-import com.google.auth.CredentialTypeForMetrics;
+import com.google.auth.oauth2.MetricsUtils.CredentialTypeForMetrics;
 import com.google.auth.ServiceAccountSigner;
 import com.google.auth.http.HttpCredentialsAdapter;
 import com.google.auth.http.HttpTransportFactory;
