+++ oauth2_http/java/com/google/auth/oauth2/MetricsUtils.java
@@ -32,7 +32,6 @@
 package com.google.auth.oauth2;
 
 import com.google.api.client.http.HttpRequest;
-import com.google.auth.CredentialTypeForMetrics;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Properties;
@@ -115,5 +114,37 @@ class MetricsUtils {
 
   static void setMetricsHeader(HttpRequest request, String metricsHeader) {
     request.getHeaders().set(MetricsUtils.API_CLIENT_HEADER, metricsHeader);
+  }
+
+  /**
+   * Defines the different types of credentials that can be used for metrics.
+   *
+   * <p>Each credential type is associated with a label that is used for reporting purposes. Add new
+   * enum constant only when corresponding configs established.
+   *
+   * <p>Credentials with type {@code CredentialTypeForMetrics.DO_NOT_SEND} is default value for
+   * credential implementations that do not set type specifically. It is not expected to send metrics.
+   *
+   * <p>
+   *
+   * @see #getLabel()
+   */
+  public enum CredentialTypeForMetrics {
+    USER_CREDENTIALS("u"),
+    SERVICE_ACCOUNT_CREDENTIALS_AT("sa"),
+    SERVICE_ACCOUNT_CREDENTIALS_JWT("jwt"),
+    VM_CREDENTIALS("mds"),
+    IMPERSONATED_CREDENTIALS("imp"),
+    DO_NOT_SEND("dns");
+
+    private final String label;
+
+    private CredentialTypeForMetrics(String label) {
+      this.label = label;
+    }
+
+    public String getLabel() {
+      return label;
+    }
   }
 }
