+++ oauth2_http/javatests/com/google/auth/oauth2/MetricsUtilsTest.java
@@ -33,7 +33,7 @@ package com.google.auth.oauth2;
 
 import static org.junit.Assert.*;
 
-import com.google.auth.CredentialTypeForMetrics;
+import com.google.auth.oauth2.MetricsUtils.CredentialTypeForMetrics;
 import com.google.auth.oauth2.MetricsUtils.RequestType;
 import java.util.regex.Pattern;
 import org.junit.Test;
